cli & main code | patch update
Version:0.0.26

+++ code/ArxFramework/config/global_config.rs
@@ -22,7 +22,7 @@ pub enum ApplicationType {
 /// e il numero massimo di thread. Può essere esteso con nuovi campi secondo necessità.
 pub struct CoreConfig {
     pub app_type: ApplicationType,
-    pub max_threads: usize,
+    pub max_threads: u8,
 }
 
 /// Configurazione della memoria per il sistema.
@@ -32,7 +32,7 @@ pub struct CoreConfig {
 pub struct MemoryConfig {
     pub pool_size: usize,    // Dimensione del pool di buffer (per PoolBased)
     pub buffer_size: usize,  // Dimensione del buffer (per Embedded)
-    pub memory_scale: f32,   // Scala per la dimensione della memoria dei modelli
+    pub memory_scale: u8,   // Scala per la dimensione della memoria dei modelli
 }
 
 /// Implementazione del valore di default per `MemoryConfig`.
@@ -43,7 +43,7 @@ impl Default for MemoryConfig {
         MemoryConfig {
             pool_size: 10,      // Valore di default: 10 buffer per il pool
             buffer_size: 1024,   // Valore di default: 1024 byte per buffer
-            memory_scale: 1.0,  // Valore di default: scala 1.0 per la dimensione della memoria
+            memory_scale: 1,  // Valore di default: scala 1.0 per la dimensione della memoria
         }
     }
 }
